+++ R/files.R
@@ -1,103 +1,84 @@
-#' Create directories if they don't already exist.
+#' Create directories if they don't already exist
 #'
 #' Given the names of (potential) directories, create the ones that do not
 #' already exist.
-#' @param dirs The name of the directories, specified via relative or
-#'   absolute paths.
+#' @param ... The names of the directories, specified via relative or absolute
+#'   paths. Duplicates are ignored.
 #' @return Invisibly, a vector with a `TRUE` for each time a directory was
-#'   actually created and a `FALSE` otherwise.
+#'   actually created and a `FALSE` otherwise. This vector is named with the
+#'   paths of the directories that were passed to the function.
 #' @examples
 #' setwd(tempdir())
-#' create_dirs(c("mydir", "yourdir"))
-#' remove_dirs(c("mydir", "yourdir"))
+#' create_dir(c("mydir", "yourdir"))
+#' remove_dir(c("mydir", "yourdir"))
 #' @export
-create_dirs <- function(dirs) {
-  created <- vapply(dirs,
-                    function(dir) {
-                      if (!dir.exists(dir)) {
-                        dir.create(dir)
-                        TRUE
-                      } else {
-                        FALSE
-                      }
-                    }
-  , logical(1))
+create_dir <- function(...) {
+  dirs <- unique(unlist(...))
+  created <- purrr::map_lgl(dirs, function(dir) {
+                                    if (!dir.exists(dir)) {
+                                      dir.create(dir)
+                                      TRUE
+                                    } else {
+                                      FALSE
+                                    }
+                                  })
+  names(created) <- dirs
+  msg <- sum(created) %>% {
+    ifelse(., paste(., ifelse(. == 1, "directory", "directories"),
+                    "created. "),
+           "")
+  }
+  n_not_created <- sum(!created)
+  if (n_not_created) {
+    msg %<>% paste0(n_not_created,
+                    ifelse(n_not_created == 1, " directory", " directories"),
+                    " not created because ",
+                    ifelse(n_not_created == 1,
+                           "it already exists.",
+                           "they already exist."))
+  }
+  message(msg)
+  invisible(created)
 }
 
-#' Remove directories.
+#' Remove directories
 #'
 #' Delete directories and all of their contents.
-#' @param dirs The names of the directories, specified via relative or absolute
+#' @param ... The names of the directories, specified via relative or absolute
 #'   paths.
 #' @return Invisibly, a logical vector with `TRUE` for each success and
-#'   `FALSE` for failures. See [base::unlink].
+#'   `FALSE` for failures.
 #' @examples
 #' setwd(tempdir())
 #' sapply(c("mydir1", "mydir2"), dir.create)
-#' remove_dirs(c("mydir1", "mydir2"))
+#' remove_dir(c("mydir1", "mydir2"))
 #' @export
-remove_dirs <- function(dirs) {
-  outcome <- !vapply(dirs, function(dir) unlink(dir, recursive = TRUE),
-                     integer(1))
-  message(sum(outcome), " directories deleted. ", sum(!outcome),
-          " failed to delete.")
+remove_dir <- function(...) {
+  dirs <- unlist(...)
+  outcome <- !as.logical(purrr::map_int(dirs, unlink, recursive = TRUE))
+  outcome %>% {
+    paste(sum(.), ifelse(sum(.) == 1, "directory", "directories"),
+          "deleted.", sum(!.), "failed to delete.")
+  } %>% message()
   invisible(outcome)
 }
 
-#' @rdname remove_dirs
+#' @rdname remove_dir
 #' @export
-dir.remove <- remove_dirs
+dir.remove <- remove_dir
 
-#' Merge tables on disk.
-#'
-#' Merge tables saved on disk as delimited files. The merging is done lengthways
-#' so they need to have the same number of columns and the same column names (if
-#' they have column names).
-#'
-#' @param files The paths to the files to merge.
-#' @param delim Delimeter used to separate values.
-#' @param out_name The path to the output file containing the merged tables.
-#' @param header Do the tables to be merged have headers?
-#' @param ... Additional arguments passed to [readr::read_delim].
-#' @examples
-#' setwd(tempdir())
-#' dir.create("MergeTablesOnDisk_test")
-#' setwd("MergeTablesOnDisk_test")
-#' tab1 <- tibble::tibble(x = 1, y = 2)
-#' tab2 <- tibble::tibble(x = 1, y = 29)
-#' mapply(readr::write_csv, list(tab1, tab2), paste0(c("tab1", "tab2"), ".csv"))
-#' merge_tables_on_disk(c("tab1.csv", "tab2.csv"), ",", "merged.csv")
-#' readr::read_csv("merged.csv")
-#' setwd("..")
-#' dir.remove("MergeTablesOnDisk_test")
-#' @export
-merge_tables_on_disk <- function(files, delim,
-                                 out_name, header = TRUE, ...) {
-  tables <- lapply(files, readr::read_delim, delim, col_names = header, ...)
-  ncs <- vapply(tables, ncol, integer(1))
-  if (!all_equal(ncs)) stop("The tables have different numbers of columns.")
-  if (header) {
-    namess <- vapply(tables, names, character(ncs[1]))
-    if (!all(apply(namess, 1, all_equal))) {
-      stop("Tables have different colnames.")
-    }
-  }
-  merged <- Reduce(rbind, tables)
-  readr::write_delim(merged, out_name, delim = delim, col_names = header)
-}
-
-MoveFile <- function(file, destination) {
+move_file <- function(file, destination) {
   # This also works for directories
   file <- normalizePath(file)  # get full path
-  file.name.base <- basename(file)
+  file_name_base <- basename(file)
   destination <- normalizePath(destination)  # remove risk of tilde use
-  new.name <- paste0(destination, "/", file.name.base)
-  file.rename(file, new.name)
+  new_name <- paste0(destination, "/", file_name_base)
+  file.rename(file, new_name)
 }
 
 #' Move files around.
 #'
-#' Move specified files into specified directories.
+#' Move specified files into specified directories
 #'
 #' If there are \eqn{n} files, there must be either \eqn{1} or \eqn{n}
 #' directories. If there is one directory, then all \eqn{n} files are moved
@@ -113,19 +94,31 @@ MoveFile <- function(file, destination) {
 #'   which to move the files.
 #' @return Invisibly, a logical vector with a `TRUE` for each time the operation
 #'   succeeded and a `FALSE` for every fail.
+#' @examples
+#' setwd(tempdir())
+#' dir.create("dir")
+#' files <- c("1litres_1.txt", "1litres_30.txt", "3litres_5.txt")
+#' file.create(files)
+#' file.move(files, "dir")
 #' @export
 move_files <- function(files, destinations) {
-  create_dirs(destinations)
-  if(length(destinations) == length(files)) {
-    outcome <- mapply(MoveFile, files, destinations)
-  } else if (length(destinations) == 1) {
-    outcome <- vapply(files, MoveFile, logical(1), destinations)
-  } else {
-    stop("the number of destinations must be equal to 1 or equal to the ",
+  if (! length(destinations) %in% (c(1, length(files)))) {
+    stop("The number of destinations must be equal to 1 or equal to the ",
          "number of files to be moved")
   }
-  message(sum(outcome), " files moved. ", sum(!outcome),
-          " failed to move.")
+  n_created_dirs <- sum(suppressMessages(create_dir(destinations)))
+  if (n_created_dirs > 0) {
+    message(n_created_dirs, " ", "director",
+            ifelse(n_created_dirs == 1, "y", "ies"),
+            " created.")
+  }
+  if(length(destinations) == length(files)) {
+    outcome <- purrr::map2_lgl(files, destinations, move_file)
+  } else {
+    outcome <- purrr::map_lgl(files, move_file, destinations)
+  }
+  message(sum(outcome), ifelse(sum(outcome) == 1, " file", " files"),
+          " moved. ", sum(!outcome), " failed.")
   invisible(outcome)
 }
 
@@ -154,7 +147,7 @@ file.move <- move_files
 #' @param pattern A regular expression. If specified, files to be renamed are
 #'   restricted to ones matching this pattern (in their name).
 #' @return A logical vector with a `TRUE` for each successful rename
-#'   (should be all TRUEs) and `FALSE`s otherwise.
+#'   (should be all `TRUE`s) and a `FALSE` otherwise.
 #'
 #' @examples
 #' setwd(tempdir())
@@ -168,8 +161,8 @@ file.move <- move_files
 #' dir.remove("NiceFileNums_test")
 #' @export
 nice_file_nums <- function(dir = ".", pattern = NA) {
-  init.dir <- getwd()
-  on.exit(setwd(init.dir))
+  init_dir <- getwd()
+  on.exit(setwd(init_dir))
   setwd(dir)
   if (is.na(pattern)) {
     lf <- list.files()
@@ -189,7 +182,7 @@ nice_file_nums <- function(dir = ".", pattern = NA) {
 #'
 #' @param dir The directory in which to perform the operation.
 #' @param pattern A regular expression. If specified, only files matching this
-#'   pattern wil be treated.
+#'   pattern will be treated.
 #' @param replacement What do you want to replace the spaces with? This
 #'   defaults to nothing, another sensible choice would be an underscore.
 #' @return A logical vector indicating which operation succeeded for each of the
@@ -207,25 +200,27 @@ nice_file_nums <- function(dir = ".", pattern = NA) {
 #' dir.remove("RemoveFileNameSpaces_test")
 #' @export
 remove_filename_spaces <- function(dir = ".", pattern = "", replacement = "") {
-  init.dir <- getwd()
-  on.exit(setwd(init.dir))
+  init_dir <- getwd()
+  on.exit(setwd(init_dir))
   setwd(dir)
   lf <- list.files(pattern = pattern)
-  new.names <- str_replace_all(lf, " ", replacement)
-  outcome <- file.rename(lf, new.names)
+  new_names <- str_replace_all(lf, " ", replacement)
+  outcome <- file.rename(lf, new_names)
   message(sum(outcome), " files renamed. ", sum(!outcome),
           " failed to rename.")
   invisible(outcome)
 }
 
-#' Replace file names with numbers.
+#' Replace file names with numbers
 #'
 #' Rename the files in the directory, replacing file names with numbers only.
+#'
 #' @param dir The directory in which to rename the files (relative or absolute
 #'   path). Defaults to current working directory.
 #' @param pattern A regular expression. If specified, only files with names
 #'   matching this pattern will be treated.
-#' @return A logical vector with a `TRUE` for each successful renaming.
+#' @return A logical vector with a `TRUE` for each successful renaming and a
+#'   `FALSE` otherwise.
 #' @examples
 #' setwd(tempdir())
 #' dir.create("RenameWithNums_test")
@@ -238,8 +233,8 @@ remove_filename_spaces <- function(dir = ".", pattern = "", replacement = "") {
 #' dir.remove("RenameWithNums_test")
 #' @export
 rename_with_nums <- function(dir = ".", pattern = NULL) {
-  init.dir <- getwd()
-  on.exit(setwd(init.dir))
+  init_dir <- getwd()
+  on.exit(setwd(init_dir))
   setwd(dir)
   lf <- list.files(pattern = pattern)
   ext <- unique(tools::file_ext(lf))
@@ -256,7 +251,7 @@ rename_with_nums <- function(dir = ".", pattern = NULL) {
   file.rename(lf, new_names)
 }
 
-#' Put files with the same unit measurements into directories.
+#' Put files with the same unit measurements into directories
 #'
 #' Say you have a number of files with "5min" in their names, number with
 #' "10min" in the names, a number with "15min" in their names and so on, and
@@ -267,7 +262,7 @@ rename_with_nums <- function(dir = ".", pattern = NULL) {
 #' [nth_number()]) before the first occurrence of the unit name. There
 #' is the option to only treat files matching a certain pattern.
 #'
-#' @param unit The unit upon which to base the categorising.
+#' @param unit The unit upon which to base the categorizing.
 #' @param pattern If set, only files with names matching this pattern will be
 #'   treated.
 #' @param dir In which directory do you want to perform this action (defaults
@@ -289,8 +284,8 @@ unitize_dirs <- function(unit, pattern = NULL, dir = ".") {
   if (!all(str_detect(lf, unit))) {
     stop(paste0("The file names must all contain the word", unit ,"."))
   }
-  up_to_first_units <- str_before_nth(lf, unit, 1)
-  nums <- vapply(up_to_first_units, nth_number, numeric(1), -1, decimals = TRUE)
+  up_to_first_units <- str_before_first(lf, unit)
+  nums <- purrr::map_dbl(up_to_first_units, last_number, decimals = TRUE)
   un <- unique(nums)
   for (i in un) {
     files <- lf[nums == i]