+++ src/strings.cpp
@@ -11,19 +11,19 @@ std::string paste_collapse(CharacterVector strings, std::string collapse) {
   }
   return out;
 }
-//' Apply paste collapse to each element of a list.
-//'
-//' This is the same as doing
-//' \code{sapply(char.list, paste, collapse = collapse)}, it's just faster.
-//'
-//' @param char_list A list of character vectors.
-//' @param collapse See \code{\link{paste}}.
-//'
-//' @return A list of character vectors.
-//'
-//' @examples
-//' paste_collapse_list_elems(list(1:3, c("a", 5, "rory")), collapse = "R")
-//' @export
+// Apply paste collapse to each element of a list.
+//
+// This is the same as doing
+// \code{sapply(char.list, paste, collapse = collapse)}, it's just faster.
+//
+// @param char_list A list of character vectors.
+// @param collapse See \code{\link{paste}}.
+//
+// @return A list of character vectors.
+//
+// @examples
+// paste_collapse_list_elems(list(1:3, c("a", 5, "rory")), collapse = "R")
+// @export
 // [[Rcpp::export]]
 CharacterVector paste_collapse_list_elems(List char_list,
                                           std::string collapse = "") {
@@ -36,15 +36,15 @@ CharacterVector paste_collapse_list_elems(List char_list,
   return(pasted);
 }
 
-//' Remove empty strings from a character list.
-//'
-//' @param char_list A list of character vectors.
-//'
-//' @return A list of character vectors.
-//'
-//' @examples
-//' str_list_remove_empties(list(c("a", "", "b"), "gg", c("", 1, "")))
-//' @export
+// Remove empty strings from a character list.
+//
+// @param char_list A list of character vectors.
+//
+// @return A list of character vectors.
+//
+// @examples
+// str_list_remove_empties(list(c("a", "", "b"), "gg", c("", 1, "")))
+// @export
 // [[Rcpp::export]]
 List str_list_remove_empties(List char_list) {
   List no_empties = clone(char_list);
@@ -62,21 +62,21 @@ List str_list_remove_empties(List char_list) {
   return(no_empties);
 }
 
-//' Get the nth element of each vector in a list of numeric or character
-//' vectors.
-//'
-//' These are faster implementations of procedures that could very easily be
-//' done with [base::sapply].
-//'
-//' @param char_list A list of character vectors.
-//' @param n The index of the element that you want from each vector.
-//'
-//' @return A list.
-//'
-//' @examples
-//' str_list_nth_elems(list(c("a", "b", "c"), c("d", "f", "a")), 2)
-//' num_list_nth_elems(list(1:5, 0:2), 4)
-//' @export
+// Get the nth element of each vector in a list of numeric or character
+// vectors.
+//
+// These are faster implementations of procedures that could very easily be
+// done with [base::sapply].
+//
+// @param char_list A list of character vectors.
+// @param n The index of the element that you want from each vector.
+//
+// @return A list.
+//
+// @examples
+// str_list_nth_elems(list(c("a", "b", "c"), c("d", "f", "a")), 2)
+// num_list_nth_elems(list(1:5, 0:2), 4)
+// @export
 // [[Rcpp::export]]
 CharacterVector str_list_nth_elems(List char_list, int n) {
   int sls = char_list.size();
@@ -90,9 +90,9 @@ CharacterVector str_list_nth_elems(List char_list, int n) {
   return(nths);
 }
 
-//' @rdname str_list_nth_elems
-//' @param num_list A list of numeric vectors.
-//' @export
+// @rdname str_list_nth_elems
+// @param num_list A list of numeric vectors.
+// @export
 // [[Rcpp::export]]
 NumericVector num_list_nth_elems(List num_list, int n) {
   int sls = num_list.size();
@@ -197,4 +197,32 @@ List interleave_char_lists(List strings1, List strings2) {
     }
   }
   return(interleaved);
+}
+
+// [[Rcpp::export]]
+IntegerVector intmat_list_nrows(List intmat_list) {
+  const std::size_t intmat_list_size = intmat_list.size();
+  IntegerVector nrows(intmat_list_size);
+  for (std::size_t i = 0; i != intmat_list_size; ++i) {
+    IntegerMatrix mat_i = as<IntegerMatrix>(intmat_list[i]);
+    nrows[i] = mat_i.nrow();
+  }
+  return nrows;
+}
+
+// [[Rcpp::export]]
+IntegerVector intmat_list_bind_nth_rows(List intmat_list, IntegerVector n) {
+  IntegerMatrix mat1 = as<IntegerMatrix>(intmat_list[0]);
+  std::size_t nc = mat1.ncol();
+  std::size_t intmat_list_size = intmat_list.size();
+  IntegerMatrix out(intmat_list_size, nc);
+  std::copy(mat1.row(n[0]).begin(), mat1.row(n[0]).end(), out.row(0).begin());
+  if (intmat_list_size > 1) {
+    for (std::size_t i = 1; i != intmat_list_size; ++i) {
+      IntegerMatrix mat_i = as<IntegerMatrix>(intmat_list[i]);
+      std::copy(mat_i.row(n[i]).begin(), mat_i.row(n[i]).end(),
+                out.row(i).begin());
+    }
+  }
+  return out;
 }